+++ tomcat/src/main/java/com/techcourse/service/UserService.java
@@ -4,8 +4,8 @@ import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.util.UUID;
 import org.apache.coyote.http11.data.HttpRequestParameter;
-import org.apache.coyote.http11.session.Session;
-import org.apache.coyote.http11.session.SessionManager;
+import org.apache.catalina.session.Session;
+import org.apache.catalina.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
