+++ tomcat/src/main/java/org/apache/catalina/session/Session.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.session;
+package org.apache.catalina.session;
 
 import java.util.HashMap;
 import java.util.Map;
